+++ src/imio/smartweb/core/contents/sections/events/content.py
@@ -89,3 +89,4 @@ class SectionEvents(Section):
     """SectionEvents class"""
 
     manage_display = True
+    show_items_date = True